Adding a ViewModel::Auction class
The goal will be to migrate view-only methods from Presenter::Auction
into here, but just making all tests pass first with delegation.
