+++ spec/models/view_model/auction_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe ViewModel::Auction, type: :model do
+end
